requirements.txt
streamlit
pandas
tavily-python
openai
requests
python-dotenv
kaggle
